NU-2361 change approach
