+++ designer/client/src/components/aiAssistant/components/AiAssistantButton.tsx
@@ -4,6 +4,7 @@ import React, { useCallback, useMemo, useRef } from "react";
 import { useTranslation } from "react-i18next";
 
 import NuIcon from "../../../assets/img/nussknacker-logo-icon.svg";
+import { convertViewportUnitToPixels } from "../../../common/convertViewportUnitToPixels";
 import { blendDarken } from "../../../containers/theme/helpers";
 import { getFeatureSettings } from "../../../reducers/selectors/settings";
 import { useAppSelector } from "../../../store/storeHelpers";
@@ -11,22 +12,6 @@ import { useWindows } from "../../../windowManager/useWindows";
 import { WindowKind } from "../../../windowManager/WindowKind";
 import DragWrapper from "./DragWrapper";
 
-function convertViewportUnitToPixels(unitString: string): number {
-    const trimmed = unitString.trim();
-    const value = parseFloat(trimmed);
-
-    if (isNaN(value)) {
-        throw new Error("Invalid input");
-    }
-
-    if (trimmed.endsWith("vw")) {
-        return (value / 100) * window.innerWidth;
-    } else if (trimmed.endsWith("vh")) {
-        return (value / 100) * window.innerHeight;
-    } else {
-        throw new Error("Unsupported unit. Use 'vw' or 'vh'.");
-    }
-}
 // TODO: use Fab
 const StyledAiAssistantButton = styled(Box)(({ theme }) => ({
     width: 75,
