+++ designer/client/src/components/toolbars/test/buttons/ScenarioTestButton.tsx
@@ -4,6 +4,7 @@ import { useTranslation } from "react-i18next";
 
 import { testScenarioWithDataRecords } from "../../../../actions/nk/displayTestResults";
 import TestingIcon from "../../../../assets/img/toolbarButtons/test.svg";
+import { convertViewportUnitToPixels } from "../../../../common/convertViewportUnitToPixels";
 import { TestCapabilityStatus } from "../../../../common/TestResultUtils";
 import {
     getTestCapabilities,
@@ -98,17 +99,17 @@ function ScenarioTestButton(props: PropsOfButton<CustomButtonTypes.scenarioTest>
                 isResizable: true,
                 isModal: true,
                 kind: WindowKind.scenarioTest,
-                minWidth: 1200,
-                minHeight: 750,
-                // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-                // @ts-ignore
-                height: "80vh",
                 meta: {
                     viewParams: {
                         Icon: TestingIcon,
                         docs,
                         markdownContent,
                     },
+                },
+                layoutData: {
+                    minWidth: 1200,
+                    minHeight: 750,
+                    height: convertViewportUnitToPixels("80vh"),
                 },
             });
         },
